requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
python-dotenv>=1.0.0
streamlit
langchain_core
langchain_openai